Add tests and rename folders
