--- TextAnalyticsFunc/SummarizeWebpage/contracts.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class WebpageText:
-    """Class representing webpage text in the database"""
-
-
