+++ TextAnalyticsFunc/host.json
@@ -11,5 +11,6 @@
   "extensionBundle": {
     "id": "Microsoft.Azure.Functions.ExtensionBundle",
     "version": "[1.*, 2.0.0)"
-  }
+  },
+  "functions": ["summarize_webpage"]
 }
